refactor: extract core extension intent contributions
